+++ src/components/film-card-full.js
@@ -1,5 +1,3 @@
-import {COMMENT_EMOJIS} from "../const.js";
-import {MONTH_NAMES} from "../const.js";
 import AbstractSmartComponent from "./abstract-smart-component.js";
 import moment from "moment";
 
@@ -11,55 +9,7 @@ const createGenresTemplate = (genres) => {
   }).join(`\n`);
 };
 
-const createEmojiImage = (name) => {
-  const image = document.createElement(`img`);
-  image.width = 55;
-  image.height = 55;
-  image.src = `images/emoji/${name}.png`;
-  image.alt = `emoji-${name}`;
-  return image;
-};
-
-const createCommentsTemplate = (comments) => {
-  return comments.map((comment) => {
-    const {text, emotion, author, date} = comment;
-    const commentDate = moment(date).format(`YYYY/MM/DD`);
-    return (
-      `<li class="film-details__comment">
-            <span class="film-details__comment-emoji">
-              <img src="images/emoji/${emotion}.png" width="55" height="55" alt="emoji-{emotion}">
-            </span>
-            <div>
-              <p class="film-details__comment-text">${text}</p>
-              <p class="film-details__comment-info">
-                <span class="film-details__comment-author">${author}</span>
-                <span class="film-details__comment-day">${commentDate}</span>
-                <button class="film-details__comment-delete">Delete</button>
-              </p>
-            </div>
-          </li>`
-    );
-  }).join(`\n`);
-};
-
-const createEmojiItemTemplate = (names) => {
-  return names.map((name) => {
-    return (
-      `<input
-        class="film-details__emoji-item visually-hidden"
-        name="comment-emoji"
-        type="radio"
-        id="emoji-${name}"
-        value="${name}"
-      >
-      <label class="film-details__emoji-label" for="emoji-${name}">
-        <img src="images/emoji/${name}.png" width="30" height="30" alt="emoji">
-      </label>`
-    );
-  }).join(`\n`);
-};
-
-const createFilmCardFullTemplate = (film, options) => {
+const createFilmCardFullTemplate = (film) => {
   const {
     poster,
     title,
@@ -67,7 +17,6 @@ const createFilmCardFullTemplate = (film, options) => {
     durationMinutes,
     genres,
     description,
-    comments,
     director,
     writers,
     actors,
@@ -79,11 +28,9 @@ const createFilmCardFullTemplate = (film, options) => {
     isFavorite,
   } = film;
 
-  const {emojiName} = options;
-
   const date = moment(productionDate).format(`D MMMM`);
   const year = moment(productionDate).format(`gggg`);
-  const durationHours = moment.utc(moment.duration(durationMinutes, "minutes").asMilliseconds()).format("H[h] mm[m]");
+  const durationHours = moment.utc(moment.duration(durationMinutes, `minutes`).asMilliseconds()).format(`H[h] mm[m]`);
 
   return (
     `<section class="film-details">
@@ -145,28 +92,6 @@ const createFilmCardFullTemplate = (film, options) => {
         <label for="favorite" class="film-details__control-label film-details__control-label--favorite">Add to favorites</label>
       </section>
     </div>
-
-    <div class="form-details__bottom-container">
-      <section class="film-details__comments-wrap">
-        <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${comments.length}</span></h3>
-
-        <ul class="film-details__comments-list">
-          ${createCommentsTemplate(comments)}
-        </ul>
-         <div class="film-details__new-comment">
-          <div for="add-emoji" class="film-details__add-emoji-label">
-          ${emojiName.length > 0 ? `<img src="images/emoji/${emojiName}.png" width="55" height="55" alt="emoji-${emojiName}">` : ``}</div>
-
-          <label class="film-details__comment-label">
-            <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment">${emojiName.length > 0 ? `Great movie!` : ``}</textarea>
-          </label>
-
-          <div class="film-details__emoji-list">
-            ${createEmojiItemTemplate(COMMENT_EMOJIS)}
-          </div>
-        </div>
-      </section>
-    </div>
   </form>
 </section>`
   );
@@ -177,41 +102,18 @@ export default class FilmCardFull extends AbstractSmartComponent {
     super();
     this._film = film;
 
-    this._emojiName = ``;
-
     this._closeButtonHandler = null;
     this._setControlButtonsChangeHandler = null;
-    this._subscribeOnEvents();
 
   }
 
   getTemplate() {
-    return createFilmCardFullTemplate(this._film, {emojiName: this._emojiName});
+    return createFilmCardFullTemplate(this._film);
   }
 
   recoveryListeners() {
     this.setCloseButtonHandler(this._closeButtonHandler);
     this.setControlButtonsChangeHandler(this._setControlButtonsChangeHandler);
-    this._subscribeOnEvents();
-  }
-
-  _subscribeOnEvents() {
-    const element = this.getElement();
-    const imageBlockElement = element.querySelector(`.film-details__add-emoji-label`);
-    const text = element.querySelector(`.film-details__comment-input`);
-
-    const rerenderWithCommentEmoji = (name) => {
-      this._emojiName = name;
-      const image = createEmojiImage(name);
-      imageBlockElement.innerHTML = ``;
-      imageBlockElement.appendChild(image);
-      text.textContent = `Great movie!`;
-    };
-
-    element.querySelector(`.film-details__emoji-list`).addEventListener(`change`, (evt) => {
-      rerenderWithCommentEmoji(evt.target.value);
-    });
-
   }
 
   setCloseButtonHandler(handler) {
