+++ src/controllers/film.js
@@ -3,6 +3,7 @@ import FilmCardFullComponent from "../components/film-card-full.js";
 import {render, remove, replace} from "../utils/render.js";
 import {FilmCardViewMode as ViewMode, ButtonID} from "../const.js";
 import {isEscKey} from "../utils/keyboard.js";
+import CommentsComponent from "../components/comments";
 
 export default class FilmController {
   constructor(container, onDataChange, onViewChange) {
@@ -19,15 +20,25 @@ export default class FilmController {
     this._onEscKeyDown = this._onEscKeyDown.bind(this);
     this._showFilmPopup = this._showFilmPopup.bind(this);
     this._hideFilmPopup = this._hideFilmPopup.bind(this);
+
+    this._emojiName = ``;
+
+    this._commentsComponent = null;
+    this._comments = null;
   }
 
   render(film) {
     this._film = film;
+
     const oldFilmCardComponent = this._filmCardComponent;
     const oldFilmCardFullComponent = this._filmCardFullComponent;
 
     this._filmCardComponent = new FilmCardComponent(film);
     this._filmCardFullComponent = new FilmCardFullComponent(film);
+
+    this._comments = this._film.comments;
+    const commentsComponent = new CommentsComponent(this._comments, this._emojiName);
+    this._commentsComponent = commentsComponent;
 
     this._filmCardComponent.setClickHandler(() => {
       this._showFilmPopup();
@@ -76,12 +87,12 @@ export default class FilmController {
     if (oldFilmCardComponent && oldFilmCardFullComponent) {
       replace(this._filmCardComponent, oldFilmCardComponent);
       replace(this._filmCardFullComponent, oldFilmCardFullComponent);
+      render(this._filmCardFullComponent.getElement(), this._commentsComponent);
     } else {
       render(this._container, this._filmCardComponent);
     }
 
   }
-
 
   getFilm() {
     return this._filmCardComponent.getFilmData();
@@ -95,15 +106,19 @@ export default class FilmController {
 
   _showFilmPopup() {
     this._onViewChange();
-    this._mode = ViewMode.POPUP;
     this._filmCardFullComponent.rerender();
+
     render(this._container, this._filmCardFullComponent);
+    render(this._filmCardFullComponent.getElement(), this._commentsComponent);
+    this._mode = ViewMode.POPUP;
+
     document.addEventListener(`keydown`, this._onEscKeyDown);
   }
 
   _hideFilmPopup() {
     this._mode = ViewMode.DEFAULT;
     remove(this._filmCardFullComponent);
+
     document.removeEventListener(`keydown`, this._onEscKeyDown);
   }
